ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
gitpython==3.1.43
confluent-kafka==2.4.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
sentence-transformers==3.0.1
chromadb==0.5.3
python-dotenv==1.0.1